drug_modelling/requirements.txt
pandas
numpy
scikit-learn
joblib
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
boto3==1.40.13
botocore==1.40.13
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.19.1
fonttools==4.58.2
holidays==0.75
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
prophet==1.1.7
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.26.0
s3transfer==0.13.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
snowflake-connector-python==3.17.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.4
threadpoolctl==3.6.0
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
xgboost==3.0.2
#pywin32==310 ## Windows-only package